+++ tests/hand_written.rs
@@ -1,6 +1,8 @@
 mod hand_written {
     mod border_and_padding;
     mod caching;
+    #[cfg(feature = "float_layout")]
+    mod float_clearance;
     mod measure;
     mod min_max_overrides;
     mod relayout;